requirements.txt
openai
crewai
crewai[tools]
langchain
streamlit
python-dotenv
pathlib